pushing fix-18

+++ Debuger.py
@@ -13,7 +13,7 @@ import logging
 import datetime
 import base64
 from os import getcwd  , path
-from Database import Database
+
 import os
 OS_Info = ''
 if os.name == 'posix':
@@ -29,10 +29,14 @@ class Debuger(object):
 
     # Constuctor
     def __init__(self):
-        self.configFilePath= basePath+'/bin/conf.txt'
+
+        self.configFilePath= getcwd()+'/bin/conf.txt'
         self.loger = logging
-        ConfigFile = open(self.configFilePath,'w+')
-        ConfigFile.close()
+        if not path.isfile(self.configFilePath):
+           file =  open(self.configFilePath,'w+')
+           file.write('debugger:0')
+           file.close()
+
         if(OS_Info == 'Windows'):
             self.loger.basicConfig(filename=getcwd() + '\\debug\\debug.log',level=logging.DEBUG)
         else:
@@ -41,13 +45,10 @@ class Debuger(object):
         self.loger.info('Logging for Date '+ str(datetime.datetime.now()).rpartition('.')[0] +"\n")
         self.isdebugerOn = False
 
-        DB = Database()
         self.Information ={}
-        self.Information['debugger'] = 0
-        info = self.getDebugConfiguration()
-        if info != None:
-            if len(info) > 0 :
-                self.Information = info[0]
+
+        self.Information = self.getDebugConfiguration()
+
 
     # Function to Log Errors
     # @param msg Message to log
@@ -78,7 +79,7 @@ class Debuger(object):
 
     # Function to turn debugging On
     def tureDebugerOn(self):
-        ConfigFile = open(self.configFilePath,'r')
+
         if self.Information['debugger'] == 1:
             self.isdebugerOn = True
         else:
@@ -93,8 +94,56 @@ class Debuger(object):
         if(self.isdebugerOn):
             return str(datetime.datetime.now()).rpartition('.')[0]
 
+    def setDebugConfiguration(self,flagOfDebug):
+        try:
+            ConfigFile = open(self.configFilePath, 'w+')
+            ConfigFile.write('debugger:'+str(flagOfDebug))
+            ConfigFile.close()
+        except Exception as ex :
+            moreInformation = {"moreInfo":'null'}
+            try:
+                if not ex[0] == None:
+                    moreInformation['LogsMore'] =str(ex[0])
+            except:
+                pass
+
+            try:
+                if not ex[1] == None:
+                    moreInformation['LogsMore1'] =str(ex[1])
+            except:
+                pass
+
+            self.tureDebugerOn()
+            self.logError('Error Reporting 36 - 42 File Database While Connecting for database information'+"\n", moreInformation)
+
     def getDebugConfiguration(self):
-        print('hello')
+        Information = {}
+        Information['debugger'] = 0
+        try:
+            ConfigFile = open(self.configFilePath, 'r+')
+            debugline  = ConfigFile.readline()
+            ConfigFile.close()
+
+            debugInfo = str(debugline).split(':')
+            Information[str(debugInfo[0])] = int(debugInfo[1])
+
+        except Exception as ex :
+            moreInformation = {"moreInfo":'null'}
+            try:
+                if not ex[0] == None:
+                    moreInformation['LogsMore'] =str(ex[0])
+            except:
+                pass
+            try:
+                if not ex[1] == None:
+                    moreInformation['LogsMore1'] =str(ex[1])
+            except:
+                pass
+
+            self.tureDebugerOn()
+            self.logError('Error Reporting 36 - 42 File Database While Connecting for database information'+"\n", moreInformation)
+
+        return Information
 
 #Deg = Debuger()
 #Deg.tureDebugerOn()
